+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/HandlerMapping.java
@@ -1,7 +1,8 @@
-package com.techcourse;
+package com.interface21.webmvc.servlet.mvc.tobe;
 
 import jakarta.servlet.http.HttpServletRequest;
 
 public interface HandlerMapping {
+    
     Object getHandler(HttpServletRequest request);
 }
